Removing User tables and mentions and instead changing them to Trustee and Contractor separately, reflection change on the endpoint as well. Added to the makefile to make easier to run
