+++ Backend/property-management/src/main/kotlin/com/example/propertymanagement/controller/TrusteeController.kt
@@ -1,0 +1,40 @@
+package com.example.propertymanagement.controller
+
+import com.example.propertymanagement.model.Trustee
+import com.example.propertymanagement.service.TrusteeService
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.*
+
+@RestController
+@RequestMapping("/api/trustee")
+class TrusteeController(private val service: TrusteeService) {
+
+    @GetMapping
+    fun getAll(): List<Trustee> = service.getAll()
+
+    @GetMapping("/{id}")
+    fun getById(@PathVariable id: Long): Trustee = service.getById(id)
+
+    data class UserDto(
+        val name: String,
+        val email: String,
+        val phone: String,
+        val apikey: String,
+    )
+
+    @PostMapping
+    fun createUser(@RequestBody userDto: UserDto): Trustee {
+        return service.addUser(userDto.name, userDto.email, userDto.phone, userDto.apikey)
+    }
+
+
+    @PutMapping("/{id}")
+    fun update(@PathVariable id: Long, @RequestBody item: Trustee): Trustee =
+        service.update(id, item)
+
+    @DeleteMapping("/{id}")
+    fun delete(@PathVariable id: Long): ResponseEntity<Void> {
+        service.delete(id)
+        return ResponseEntity.noContent().build()
+    }
+}
