+++ Backend/property-management/src/main/kotlin/com/example/propertymanagement/repository/UserRepository.kt
@@ -1,6 +1,6 @@
 package com.example.propertymanagement.repository
 
-import com.example.propertymanagement.model.User
+import com.example.propertymanagement.model.Trustee
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UserRepository : JpaRepository<User, Long>
+interface TrusteeRepository : JpaRepository<Trustee, Long>
